feat: w8_JWT session

+++ Django/posts/views.py
@@ -175,12 +175,18 @@ from rest_framework.response import Response
 from rest_framework import status
 from django.http import Http404
 from django.shortcuts import get_object_or_404
+from rest_framework.permissions import IsAuthenticatedOrReadOnly # jwt 세션
+from datetime import datetime
+from rest_framework import permissions
+from config.permissions import IsNotCurfewTime, IsOwnerOrReadOnly
 
 class PostList(APIView):
+    permission_classes = [IsNotCurfewTime, IsAuthenticatedOrReadOnly]
+
     def post(self, request, format=None):
         serializer = PostSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(writer=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
@@ -190,13 +196,21 @@ class PostList(APIView):
         return Response(serializer.data)
     
 class PostDetail(APIView):
+    permission_classes = [IsNotCurfewTime, IsOwnerOrReadOnly]
+
     def get(self, request, post_id):
         post = get_object_or_404(Post, id=post_id)
+
+        self.check_object_permissions(request, post)
+
         serializer = PostSerializer(post)
         return Response(serializer.data)
     
     def put(self, request, post_id):
         post = get_object_or_404(Post, id=post_id)
+
+        self.check_object_permissions(request, post)
+        
         serializer = PostSerializer(post, data=request.data)
         if serializer.is_valid(): # update이니까 유효성 검사 필요
             serializer.save()
@@ -205,6 +219,9 @@ class PostDetail(APIView):
     
     def delete(self, request, post_id):
         post = get_object_or_404(Post, id=post_id)
+
+        self.check_object_permissions(request, post)
+
         post.delete()
         return Response(
             {
@@ -215,8 +232,10 @@ class PostDetail(APIView):
         )
 
 class CommentList(APIView):
+    permission_classes = [IsNotCurfewTime]
+
     def get(self, request, post_id): # post_id를 URL에서 받아와서 해당 게시글의 댓글 리스트를 조회하는 로직
-        comments = Comment.objects.f, cilter(post_id=post_id)
+        comments = Comment.objects.filter(post_id=post_id)
         serializer = CommentSerializer(comments, many=True)
         return Response(serializer.data)
 
@@ -228,8 +247,13 @@ class CommentList(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 class CommentDetail(APIView):
+    permission_classes = [IsNotCurfewTime, IsOwnerOrReadOnly]
+
     def delete(self, request, post_id, comment_id): # comment_id를 URL에서 받아와서 해당 댓글을 삭제하는 로직
         comment = get_object_or_404(Comment, id=comment_id)
+
+        self.check_object_permissions(request, comment)
+
         comment.delete()
         return Response(
             {"message": "댓글이 성공적으로 삭제되었습니다.",
